LayoutTests/http/tests/websocket/tests/hybi/contentextensions/block.html

<script>

if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

function endTest() {
    if (window.testRunner)
        testRunner.notifyDone();
}

var ws = new WebSocket("ws://localhost:8880/websocket/tests/hybi/contentextensions/resources/echo");

ws.onopen = function() { 
    console.log("onopen");
    ws.send("sent"); 
}
ws.onmessage = function(message) { 
    console.log("onmessage " + message.data);
}
ws.onclose = function() {
    console.log("onclose");
    endTest();
}
ws.onerror = function(ev) { 
    console.log("onerror"); 
    endTest(); 
}

setTimeout(function() { 
    console.log("timeout"); 
    endTest();
}, 3000);

</script>
